+++ packages/component-library/src/components/aladin/aladin-lite.d.ts
@@ -1,0 +1,1 @@
+declare module "@cquiroz/aladin-lite/lib/js/A";